+++ src/joedb/journal/Encoded_File.h
@@ -7,15 +7,52 @@
 namespace joedb
 {
  ////////////////////////////////////////////////////////////////////////////
- class Encoded_File: public Generic_File
+ class Readonly_Encoded_File: public Generic_File
  ////////////////////////////////////////////////////////////////////////////
  {
   private:
-   Codec &codec;
-   encoded_file::Generic_File_Database &db;
+   encoded_file::Database &db;
+   Blob_Reader &blob_reader;
 
    std::vector<char> read_buffer;
    encoded_file::id_of_buffer decoded_buffer;
+
+  protected:
+   Codec &codec;
+
+   size_t pread
+   (
+    char * const buffer,
+    const size_t size,
+    const int64_t offset
+   )
+   override;
+
+   Readonly_Encoded_File
+   (
+    Codec &codec,
+    encoded_file::Database &db,
+    Blob_Reader &blob_reader,
+    Open_Mode mode
+   );
+
+  public:
+   Readonly_Encoded_File
+   (
+    Codec &codec,
+    encoded_file::Database &db,
+    Blob_Reader &blob_reader
+   );
+
+   int64_t get_size() const override;
+ };
+
+ ////////////////////////////////////////////////////////////////////////////
+ class Encoded_File: public Readonly_Encoded_File
+ ////////////////////////////////////////////////////////////////////////////
+ {
+  private:
+   encoded_file::Generic_File_Database &db;
 
    static constexpr size_t write_buffer_total_size = 1 << 20;
    std::vector<char> write_buffer;