+++ doc/source/history.rst
@@ -5,6 +5,7 @@ History
 
   - ``Encoded_File`` supports on-the-fly coding or decoding of data. Custom
     codecs can be used. ``Brotli_Codec`` is provided for compression.
+  - ``Readonly_Client`` is now also a ``Blob_Reader``
   - Fix some potential resource leaks when throwing from constructors in
     ``Posix_File``, ``ssh::Session``, ``ssh::SFTP``, ``ssh::Forward_Channel``.
     ``CURL_File``, ``Windows_File``.